:bug: Allow construction with a mongist connection
Currently, when calling `mongoist()` with an existing mongoist
connection, the database fails to initialize properly resulting in
undefined errors when calling database or collection methods.

I added an affordance for checking if the `connectionString` is an
instance of `Database` which means we can call the connect function
on that database to initialize this mongoist instance.