Fix writeConcern mutation on insert
Some versions of mongodb mutate the options value provided to the insert
call, and previously a reference to the global `writeOpts` constant was
passed into insert calls which could then be mutated and for other
calls. This updates the collection class to always copy the values, not
pass the object in directly.