Add `addCursorFlag` method to cursors
Currently, there's no way to pass cursor flags to the mongodb native
cursor such as `noCursorTimeout`, and adding the `addCursorFlag`
function allows options to be set on the native cursor object. This uses
a separate `_flags` object stored on the cursor since the flag options
do not need to be passed to count, or size methods.