feat(connect): support promise for connectionString (#41)
There are cases where it's useful to have a synchronous reference to the Database, and where the
connection information is not yet available - particularly in test cases where you may be launching
a temporary mongo database using mongo-memory-server. To better support this use-case, the Database
class can accept and store the connectionString parameter as a Promise, and resolve it to its value
in the connect handler which is already well-positioned to wait for asynchronous resolution.