chore(ci)!: remove support for node<14
Some libraries require higher versions of node to run and given that
node 14 is 2 years old and the oldest lts version, we shouldn't support
earlier versions. This simplifies development and CI without too much
risk of backward compatibility. This is a major version change but
should not add risk to projects >=14.

Also, update eslint to the latest version.