chore(deps): update parser and linters (#69)
* chore(deps): update parser and linters

* chore(ci): remove intentionally invalid json file

* chore(ci)!: remove support for node<14

Some libraries require higher versions of node to run and given that
node 14 is 2 years old and the oldest lts version, we shouldn't support
earlier versions. This simplifies development and CI without too much
risk of backward compatibility. This is a major version change but
should not add risk to projects >=14.

Also, update eslint to the latest version.

* test: update tested versions to install on latest

In development, node >= 16 is required to use the correct lockfile
verision

* chore(ci): set ci install version

* chore(ci): lint independently

Linting doesn't need to be repeated on each node version so
run it only once and before creating test jobs.

* chore(ci): resolve syntax error

* chore(ci): remove mongodb in linting

* chore(ci): fix npm command

* chore(deps-dev): update mocha and chai to fix vulnerability (#88)

Co-authored-by: Garret Meier <garret@mixmax.com>